added a first sound track
Unity is being an ass again with the version and i don't want fuck up anything in the code so i started doing the music/sounds. I've added the first track that i did on 1BIT DRAGON software, let me know what do you guys think and if it's any good. i wouldn't use it in like a dungeon, but i think it would feel well when in town with 0 insanity or in like a main menu screen